feat: allowed member to remove himself

+++ app/services/company_service/service.py
@@ -267,9 +267,10 @@ class CompanyService:
         await self._company_action_repository.delete(company_id, user_id, CompanyActionType.INVITATION)
 
     async def remove_member(self, company_id: UUID, user_id: UUID, current_user: UserDetail) -> None:
-        company = await self._company_exists_and_user_has_permission(
-            company_id, current_user, self._user_has_edit_permission
-        )
+        company = await self.check_company_exists(company_id)
+        if current_user.id != user_id and not self._user_has_edit_permission(company, current_user):
+            raise CompanyPermissionException()
+            
         if company.owner_id == user_id:
             raise CompanyActionException('Owner cannot be removed from company')
         await self._company_action_repository.delete(company_id, user_id, CompanyActionType.MEMBERSHIP)
